# api-hub/openapi.yaml
openapi: 3.1.0
info:
  description: "API specification for Open API Hub project. This API is designed to\
    \ manage users, projects, project members, API specifications, and project credentials."
  title: Open API Hub API
  version: 1.0.0
servers:
- description: Development server
  url: http://localhost:8000
...
paths:
  /users:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/User'
                type: array
          description: List of users
        "404":
          description: Users not found
      summary: Get all users
      tags:
      - Users
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: User created successfully
        "400":
          description: Invalid input
      summary: Create new user
      tags:
      - Users
  /users/{user_id}:
    delete:
      parameters:
      - description: The ID of the user to delete
        explode: false
        in: path
        name: user_id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "204":
          description: User deleted successfully
        "404":
          description: User not found
      summary: Delete user
      tags:
      - Users
    get:
      parameters:
      - description: The ID of the user to retrieve
        explode: false
        in: path
        name: user_id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: User details
        "404":
          description: User not found
      summary: Get user by ID
      tags:
      - Users
    put:
      parameters:
      - description: The ID of the user to update
        explode: false
        in: path
        name: user_id
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: User updated successfully
        "400":
          description: Invalid input
      summary: Update user
      tags:
      - Users
  /projects:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Project'
                type: array
          description: List of projects
        "404":
          description: Projects not found
      summary: Get all projects
      tags:
      - Projects
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: Project created successfully
        "400":
          description: Invalid input
      summary: Create new project
      tags:
      - Projects
  /projects/{project_id}:
    delete:
      parameters:
      - description: The ID of the project to delete
        explode: false
        in: path
        name: project_id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "204":
          description: Project deleted successfully
        "404":
          description: Project not found
      summary: Delete project
      tags:
      - Projects
    get:
      parameters:
      - description: The ID of the project to retrieve
        explode: false
        in: path
        name: project_id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: Project details
        "404":
          description: Project not found
      summary: Get project by ID
      tags:
      - Projects
    put:
      parameters:
      - description: The ID of the project to update
        explode: false
        in: path
        name: project_id
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: Project updated successfully
        "400":
          description: Invalid input
      summary: Update project
      tags:
      - Projects
  /project_members:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ProjectMember'
                type: array
          description: List of project members
        "404":
          description: Project members not found
      summary: Get all project members
      tags:
      - Project Members
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectMember'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectMember'
          description: Project member added successfully
        "400":
          description: Invalid input
      summary: Add new project member
      tags:
      - Project Members
  /project_members/{project_member_id}:
    delete:
      parameters:
      - description: The ID of the project member to delete
        explode: false
        in: path
        name: project_member_id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "204":
          description: Project member deleted successfully
        "404":
          description: Project member not found
      summary: Delete project member
      tags:
      - Project Members
    get:
      parameters:
      - description: The ID of the project member to retrieve
        explode: false
        in: path
        name: project_member_id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectMember'
          description: Project member details
        "404":
          description: Project member not found
      summary: Get project member by ID
      tags:
      - Project Members
    put:
      parameters:
      - description: The ID of the project member to update
        explode: false
        in: path
        name: project_member_id
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectMember'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectMember'
          description: Project member updated successfully
        "400":
          description: Invalid input
      summary: Update project member
      tags:
      - Project Members
...
components:
  schemas:
    User:
      example:
        password: password
        full_name: John Doe
        created_at: 2022-01-01T00:00:00Z
        id: 1
        email: user@example.com
      properties:
        id:
          description: The unique identifier for the user.
          example: 1
          title: id
          type: integer
        email:
          description: The email address of the user.
          example: user@example.com
          title: email
          type: string
        password:
          description: The password of the user.
          example: password
          title: password
          type: string
        full_name:
          description: The full name of the user.
          example: John Doe
          title: full_name
          type: string
        created_at:
          description: The date and time when the user was created.
          example: 2022-01-01T00:00:00Z
          format: date-time
          title: created_at
          type: string
      title: User
    Project:
      example:
        is_archived: false
        name: Example Project
        description: This is an example project.
        created_at: 2022-01-01T00:00:00Z
        id: 1
        created_by: 1
      properties:
        id:
          description: The unique identifier for the project.
          example: 1
          title: id
          type: integer
        name:
          description: The name of the project.
          example: Example Project
          title: name
          type: string
        description:
          description: The description of the project.
          example: This is an example project.
          title: description
          type: string
        is_archived:
          description: Indicates whether the project is archived.
          example: false
          title: is_archived
          type: boolean
        created_by:
          description: The unique identifier of the user who created the project.
          example: 1
          title: created_by
          type: integer
        created_at:
          description: The date and time when the project was created.
          example: 2022-01-01T00:00:00Z
          format: date-time
          title: created_at
          type: string
      title: Project
    ProjectMember:
      example:
        project_id: 1
        user_id: 1
        member_role: admin
        created_at: 2022-01-01T00:00:00Z
        id: 1
      properties:
        id:
          description: The unique identifier for the project member.
          example: 1
          title: id
          type: integer
        project_id:
          description: The unique identifier of the project to which the member belongs.
          example: 1
          title: project_id
          type: integer
        user_id:
          description: The unique identifier of the user who is a member of the project.
          example: 1
          title: user_id
          type: integer
        member_role:
          description: The role of the member in the project.
          example: admin
          title: member_role
          type: string
        created_at:
          description: The date and time when the project member was created.
          example: 2022-01-01T00:00:00Z
          format: date-time
          title: created_at
          type: string
      title: ProjectMember
...
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http